Rename --diskID into --deviceID Now the deviceID can contains following device id types which can be evaluated by lsblk command: SERIAL the SCSI serial ID of a device PTUUID the UUID of the partition table PARTUUID the UUID of the partition UUID the UUID of the filesystem
New command check-auto-unlock:
The reason of this new command ist that the udev-rule delivers all UUIDs.
Befor the cryptctl2-auto-unlock@ service will be started the cryptctl2-auto-unlock.sh checks
which UUID can be used for unlocking the device.

Now AutoEncryption was implemented and a new paramenter fileSystem introduced.
If it is given after encoding the device this will be formatted with the given filesytem.